Updated a lot of documentation, added a new get-started tutorial.
