+++ opendc-experiments/opendc-experiments-base/src/main/kotlin/org/opendc/experiments/base/experiment/Scenario.kt
@@ -20,15 +20,15 @@
  * SOFTWARE.
  */
 
-package org.opendc.experiments.base.scenario
+package org.opendc.experiments.base.experiment
 
 import org.opendc.compute.simulator.telemetry.parquet.ComputeExportConfig
-import org.opendc.experiments.base.scenario.specs.AllocationPolicySpec
-import org.opendc.experiments.base.scenario.specs.CheckpointModelSpec
-import org.opendc.experiments.base.scenario.specs.ExportModelSpec
-import org.opendc.experiments.base.scenario.specs.FailureModelSpec
-import org.opendc.experiments.base.scenario.specs.ScenarioTopologySpec
-import org.opendc.experiments.base.scenario.specs.WorkloadSpec
+import org.opendc.experiments.base.experiment.specs.AllocationPolicySpec
+import org.opendc.experiments.base.experiment.specs.CheckpointModelSpec
+import org.opendc.experiments.base.experiment.specs.ExportModelSpec
+import org.opendc.experiments.base.experiment.specs.FailureModelSpec
+import org.opendc.experiments.base.experiment.specs.ScenarioTopologySpec
+import org.opendc.experiments.base.experiment.specs.WorkloadSpec
 
 /**
  * A data class representing a scenario for a set of experiments.
