+++ opendc-experiments/opendc-experiments-base/src/main/kotlin/org/opendc/experiments/base/experiment/specs/AllocationPolicySpec.kt
@@ -20,34 +20,21 @@
  * SOFTWARE.
  */
 
-package org.opendc.experiments.base.scenario
+package org.opendc.experiments.base.experiment.specs
 
-import kotlinx.serialization.ExperimentalSerializationApi
-import kotlinx.serialization.json.Json
-import kotlinx.serialization.json.decodeFromStream
-import org.opendc.compute.simulator.telemetry.parquet.ComputeExportConfig
-import org.opendc.experiments.base.scenario.specs.ExperimentSpec
-import java.io.File
-import java.io.InputStream
-import java.nio.file.Path
-import kotlin.io.path.inputStream
+import kotlinx.serialization.Serializable
+import org.opendc.compute.simulator.scheduler.ComputeSchedulerEnum
 
-public class ExperimentReader {
-    private val jsonReader = Json
-
-    public fun read(file: File): ExperimentSpec = read(file.inputStream())
-
-    public fun read(path: Path): ExperimentSpec = read(path.inputStream())
-
-    /**
-     * Read the specified [input].
-     */
-    @OptIn(ExperimentalSerializationApi::class)
-    public fun read(input: InputStream): ExperimentSpec {
-        // Loads the default parquet output fields,
-        // so that they can be deserialized
-        ComputeExportConfig.loadDfltColumns()
-
-        return jsonReader.decodeFromStream<ExperimentSpec>(input)
-    }
+/**
+ * specification describing how tasks are allocated
+ *
+ * @property policyType
+ *
+ * TODO: expand with more variables such as allowed over-subscription
+ */
+@Serializable
+public data class AllocationPolicySpec(
+    val policyType: ComputeSchedulerEnum = ComputeSchedulerEnum.Mem,
+) {
+    public val name: String = policyType.toString()
 }
