+++ opendc-experiments/opendc-experiments-base/src/main/kotlin/org/opendc/experiments/base/experiment/specs/ExportModelSpec.kt
@@ -20,22 +20,20 @@
  * SOFTWARE.
  */
 
-package org.opendc.experiments.base.scenario.specs
+package org.opendc.experiments.base.experiment.specs
 
 import kotlinx.serialization.Serializable
-import org.opendc.compute.simulator.telemetry.parquet.ComputeExportConfig
 
+/**
+ * specification describing how the results should be exported
+ *
+ * @property exportInterval The interval of exporting results in s. Should be higher than 0.0
+ */
 @Serializable
-public data class ScenarioSpec(
-    var id: Int = -1,
-    var name: String = "",
-    val outputFolder: String = "output",
-    val computeExportConfig: ComputeExportConfig,
-    val topology: ScenarioTopologySpec,
-    val workload: WorkloadSpec,
-    val allocationPolicy: AllocationPolicySpec = AllocationPolicySpec(),
-    val exportModel: ExportModelSpec = ExportModelSpec(),
-    val failureModel: FailureModelSpec? = null,
-    val checkpointModel: CheckpointModelSpec? = null,
-    val maxNumFailures: Int = 10,
-)
+public data class ExportModelSpec(
+    val exportInterval: Long = 5 * 60,
+) {
+    init {
+        require(exportInterval > 0) { "The Export interval has to be higher than 0" }
+    }
+}
