+++ opendc-experiments/opendc-experiments-base/src/main/kotlin/org/opendc/experiments/base/runner/ExperimentRunner.kt
@@ -22,7 +22,7 @@
 
 package org.opendc.experiments.base.runner
 
-import org.opendc.experiments.base.scenario.Scenario
+import org.opendc.experiments.base.experiment.Scenario
 import java.util.concurrent.ForkJoinPool
 
 /**
