+++ site/docs/getting-started/documents/experiments/simple_experiment.json
@@ -1,0 +1,13 @@
+{
+    "name": "simple",
+    "topologies": [{
+        "pathToFile": "topologies/small.json"
+    },
+    {
+        "pathToFile": "topologies/big.json"
+    }],
+    "workloads": [{
+        "pathToFile": "traces/bitbrains-small",
+        "type": "ComputeWorkload"
+    }]
+}
