+++ site/docs/getting-started/documents/topologies/big.json
@@ -1,0 +1,59 @@
+{
+    "clusters":
+    [
+        {
+            "name": "C01",
+            "hosts" :
+            [
+                {
+                    "name": "H01",
+                    "cpu":
+                    {
+                        "coreCount": 32,
+                        "coreSpeed": 3200
+                    },
+                    "memory": {
+                        "memorySize": 256000
+                    }
+                }
+            ]
+        },
+        {
+            "name": "C02",
+            "hosts" :
+            [
+                {
+                    "name": "H02",
+                    "count": 6,
+                    "cpu":
+                    {
+                        "coreCount": 8,
+                        "coreSpeed": 2930
+                    },
+                    "memory": {
+                        "memorySize": 64000
+                    }
+                }
+            ]
+        },
+        {
+            "name": "C03",
+            "hosts" :
+            [
+                {
+                    "name": "H03",
+                    "count": 2,
+                    "cpu":
+                    {
+                        "coreCount": 16,
+                        "coreSpeed": 3200
+                    },
+                    "memory": {
+                        "memorySize": 128000
+                    }
+                }
+            ]
+        }
+    ]
+}
+
