+++ site/old_tutorials/_category_.json
@@ -1,0 +1,8 @@
+{
+  "label": "Getting Started",
+  "position": 2,
+  "link": {
+    "type": "generated-index",
+    "description": "10 minutes to learn the most important concepts of OpenDC."
+  }
+}
